+++ src/back/dao/PrintBoardDAO.java
@@ -2,9 +2,6 @@ package back.dao;
 
 import back.BoardDTO;
 import back.UserDTO;
-import back.request.board.Post_Board_Request;
-import back.response.board.Board_Info_More_Response;
-import back.response.mypage.My_Board_Info_More_Response;
 import database.DBConnector;
 import back.response.board.Board_Info_Response;
 
@@ -15,12 +12,10 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class BoardDAO {
+public class PrintBoardDAO {
     Connection conn = null;
     PreparedStatement pt = null;
     ResultSet rs = null;
-
-    // 이 코드 기준으로 밑에 코드들 다 바꿔야 됨 ( 고민재 할 일 )
 
     // 게시판 출력
     public List<Board_Info_Response> printBoard(String region, String category) {
@@ -168,124 +163,5 @@ public class BoardDAO {
         System.out.println("내가 참여한 공구 내역 출력 완료.");
 
         return data;
-    }
-
-    // 게시글 자세히 보기 (메인페이지)
-    public Board_Info_More_Response readMorePost(int selectRow, String uuid) {
-        selectRow++;
-
-        String selectSQL = "SELECT * FROM boardView WHERE num = ?;";
-        String nickNameSQL = "SELECT nickName FROM user WHERE uuid = ?;";
-        String updateSQL = "UPDATE board SET view = view + 1 WHERE boardID = ?;";
-
-        try {
-            conn = DBConnector.getConnection();
-            pt = conn.prepareStatement(selectSQL);
-            pt.setInt(1, selectRow);
-            rs = pt.executeQuery();
-            pt = conn.prepareStatement(nickNameSQL);
-            rs.next();
-            pt.setString(1, rs.getString("uuid"));
-            ResultSet rs1 = pt.executeQuery();
-
-            if (rs1.next()) {
-                String peoplenum = rs.getInt("nowPeopleNum") + "/" + rs.getString("peopleNum");
-
-                Board_Info_More_Response boardInfoMoreResponse = new Board_Info_More_Response(
-                        rs.getInt("boardId"),
-                        rs.getString("title"),
-                        rs.getString("region"),
-                        rs.getString("category"),
-                        rs1.getString(1),
-                        peoplenum,
-                        rs.getString("content"),
-                        rs.getInt("view") + 1,
-                        rs.getString("uuid").equals(uuid),
-                        rs.getInt("chatPort")
-                );
-
-                pt = conn.prepareStatement(updateSQL);
-                pt.setInt(1, boardInfoMoreResponse.boardId());
-                pt.execute();
-
-                rs.close();
-                rs1.close();
-                pt.close();
-                conn.close();
-
-                return boardInfoMoreResponse;
-            }
-        } catch (Exception exception) {
-            exception.printStackTrace();
-        }
-        return null;
-    }
-
-    // 내가 쓴 글 자세히 보기 (마이페이지)
-    public My_Board_Info_More_Response readMoreMyPost(int selectRow) {
-        selectRow++;
-        String selectSQL = "SELECT * FROM boardView WHERE num = ?";
-        String updateSQL = "UPDATE board SET view = view + 1 WHERE boardID = ?;";
-        try {
-            conn = DBConnector.getConnection();
-            pt = conn.prepareStatement(selectSQL);
-            pt.setInt(1, selectRow);
-            rs = pt.executeQuery();
-            if (rs.next()) {
-                String peoplenum = rs.getInt("nowPeopleNum") +"/"+ rs.getString("peopleNum");
-
-                My_Board_Info_More_Response myBoardInfoMoreResponse = new My_Board_Info_More_Response(
-                        rs.getInt("boardId"),
-                        rs.getString("title"),
-                        rs.getString("region"),
-                        rs.getString("category"),
-                        peoplenum,
-                        rs.getString("content"),
-                        rs.getInt("view" + 1)
-                );
-                pt = conn.prepareStatement(updateSQL);
-                pt.setInt(1, myBoardInfoMoreResponse.boardId());
-                pt.execute();
-                System.out.println("자세히 보기 성공.");
-
-                rs.close();
-                pt.close();
-                conn.close();
-
-                return myBoardInfoMoreResponse;
-            }
-            rs.close();
-            pt.close();
-            conn.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    public void posting(Post_Board_Request Post_BoardInfo, int port) { // <- 이 port 정보도 게시글 테이블에 포함 해줘야돼
-        conn = DBConnector.getConnection();
-        String insertSQL = "INSERT INTO board(title, region, category, peopleNum, content, uuid, view, nowPeopleNum, chatPort) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
-        try {
-            pt = conn.prepareStatement(insertSQL);
-            pt.setString(1, Post_BoardInfo.title());
-            pt.setString(2, Post_BoardInfo.region());
-            pt.setString(3, Post_BoardInfo.category());
-            pt.setString(4, Post_BoardInfo.peopleNum());
-            pt.setString(5, Post_BoardInfo.content());
-            pt.setString(6, Post_BoardInfo.uuid());
-            pt.setInt(7, 0);
-            pt.setInt(8, 1);
-            pt.setInt(9, port);
-
-            if (!pt.execute()) {
-                System.out.println("게시 성공.");
-            }
-            pt.close();
-            conn.close();
-        } catch (Exception e) {
-            System.out.println("게시 실패.");
-            e.printStackTrace();
-        }
     }
 }
